[project @ Added ARRAY-EQUAL and supporting functions.]
Added a function, ARRAY-EQUAL, that will descend into 2 arrays and
checks each element for equality. Necessited 2 supporting
functions. The first function, DIMMENSIONS-EQUAL, returns true if the
arrays are of equal rank and equal dimension in each rank. The other
supporting function, ELEMENT-EQUAL, performs the actual work of the
testing each element of each array for equality.