[project @ Replaced the optional EPSILON in ARRAY-EQUAL with a TEST keyword.]
Instead of specifying an epsilon, it makes more sense to pass
ARRAY-EQUAL a test function. This was necessitated by the
implementation of SIGNIFICANT-FIGURES-EQUAL. In the process of
updating ARRAY-EQUAL, realized that DIMENSIONS-EQUAL was performing
redundant tests and could simply be replaced with EQUAL.

This change was propagated to the ELEMENT-EQUAL support function.