[project @ Corrected bugs in SIGNIFICANT-FIGURES-EQUAL related to values of zero.]
The SIGNIFICANT-FIGURES-EQUAL and NORMALIZE-FLOAT functions failed for
values of or near zero. The 2 primary changes in this patch corrected
those functions. In the process, it was recognized that a
multiplication using the value 0.1 was being performed in
NORMALIZE-FLOAT instead of dividing by 10.0. Repeatedly multiplying a
number by 0.1, which cannot be exactly represented by a floating point
number, caused an error to accumulate in the number. Dividing by 10.0
eliminates this error at the expense of computational
performance. Unless there is a compelling reason, always choose
correctness over performance. Finally, another convenience function,
2-SIGFIG-EQUAL, was added.