[project @ Organized the code into separate files.]
Moved the package definition to defpackage.lisp and the floating point
related functions to floating-point.lisp. This was motivated by the
desire to track any changes made to the original LISP-UNIT by the
author. By separating additional functionality into other files, it
will be easier to identify differences in lisp-unit.lisp.