Implemented diagnostic functions for numbers, sequences and arrays.
Two types of diagnostic functions are implemented for numbers. The
first type returns the roundoff error between 2 numbers. The second
type returns the roundoff error as a multiple of the appropriate
precision float-epsilon value.

For sequences and arrays, the diagnostic functions return a list of
the elements that are not equal, noting the element indices, values
and numeric error. Each function has 2 keywords. First, a :TEST
keyword to specify the equality test, default to NUMBER-EQUAL. Second,
an :ERROR-FUNCTION keyword specifying which type of error should be
reported, defaulting to #'NUMBER-ERROR.

NOTE: This is an initial implementation and subject to change.