Documentation for the floating point extensions.
Updated the documentation to include the floating point predicates,
assertions and diagnostic functions. Performed a trivial reformatting
of the HTML for better readability in emacs. Also, added an xml tag so
that the document will work in nxml-mode. It would have been
preferable to use nXhtml-mode, but that only works with emacs 22 and
RHEL only packages emacs 21.