Bug fix for NORMALIZE-FLOAT, SEQUENCE-ERROR and ARRAY-ERROR.
The biggest errors were in the iteration forms in SEQUENCE-ERROR and
ARRAY-ERROR. Originally, a DOTIMES was used with an LET form as the
body. The current form was a translation to a DO* iteration, but the
return logic was flawed. The return logic has been corrected by going
back to a DOTIMES iteration as the body of LET form and using SETF to
set the variables on each iteration. It is probably time to learn to
use LOOP.

I re-introduced the exactly equal to 10.0 when I corrected
NORMALIZE-FLOAT to return a significand in the range [0,1). This was
corrected and NORMALIZE-FLOAT correctly handles values of 10.0, again.