Expanded the predicate and assertion interface.
Expanded the testing interface with RATIONAL-EQUAL, FLOAT-EQUAL and
NORM-EQUAL. These are implemented as generic functions that operate on
atoms, sequences and arrays. In the case of sequences and arrays,
RATIONAL-EQUAL and FLOAT-EQUAL perform and element-wise comparison for
equality. NORM-EQUAL performs the test for equality on the relative
error norm of the sequence or array.