Remove the lexical environment variable from assert-expands.
Could not figure out a way to make this work on both LispWorks and SBCL.
It wasn't even clear that it was being handled correctly, anyway. In the
process, it was found the EQUAL was being used to compare the expansion
and the expected expansion. This should be replaced with the
expansion-equal function so that GENSYMS are handled correctly,
for example.