introduce a `*keep-passing-asserts*` flag to toggle keeping details about passing asserts.
When non-nil, passing test assertions will be collected as objects and
accessible in test-result objects.  When nil, only the type of the
passing assertion will be collected, saving significant memory.

The default is T, which can consume lots of memory for large numbers
of asserts (in my case, sbcl heap exhaustion with 95K assertions).

refs #8